refactor sharemenu
