+++ packages/client-core/src/components/Settings/ShareSpaceScreen.tsx
@@ -24,38 +24,15 @@ Infinite Reality Engine. All Rights Reserved.
 */
 
 import { QRCodeSVG } from 'qrcode.react'
-import React, { useEffect, useState } from 'react'
+import React from 'react'
+import { useShareMenu } from '../../hooks/useShareMenu'
 
 interface ShareSpaceScreenProps {
   navigateTo: (screen: string) => void
 }
 
 const ShareSpaceScreen: React.FC<ShareSpaceScreenProps> = () => {
-  const [shareLink, setShareLink] = useState('')
-  const [showCopyNotification, setShowCopyNotification] = useState(false)
-
-  useEffect(() => {
-    // Set the share link to the current URL
-    setShareLink(window.location.href)
-  }, [])
-
-  const copyLinkToClipboard = () => {
-    navigator.clipboard.writeText(shareLink)
-    setShowCopyNotification(true)
-    setTimeout(() => {
-      setShowCopyNotification(false)
-    }, 3000)
-  }
-
-  const shareToMetaQuest = () => {
-    // Implement Meta Quest sharing functionality
-    console.log('Sharing to Meta Quest')
-  }
-
-  const shareByEmailOrPhone = () => {
-    // Implement email/phone sharing functionality
-    console.log('Opening email/phone sharing dialog')
-  }
+  const { shareLink, copyLinkToClipboard } = useShareMenu()
 
   return (
     <div className="xs:gap-6 flex h-full flex-col items-center justify-between p-4 md:flex-row md:items-start md:justify-center md:gap-5">
@@ -75,27 +52,14 @@ const ShareSpaceScreen: React.FC<ShareSpaceScreenProps> = () => {
           Copy Direct Link
         </button>
 
-        <button
-          onClick={shareToMetaQuest}
-          className="w-full rounded-full bg-white/20 py-3 text-center text-white hover:bg-white/30"
-        >
+        <button className="w-full rounded-full bg-white/20 py-3 text-center text-white hover:bg-white/30">
           Share to Meta Quest
         </button>
 
-        <button
-          onClick={shareByEmailOrPhone}
-          className="w-full rounded-full bg-white/20 py-3 text-center text-white hover:bg-white/30"
-        >
+        <button className="w-full rounded-full bg-white/20 py-3 text-center text-white hover:bg-white/30">
           Share by email or phone
         </button>
       </div>
-
-      {/* Copy Notification */}
-      {showCopyNotification && (
-        <div className="fixed bottom-20 left-1/2 -translate-x-1/2 rounded-full bg-gray-800/90 px-6 py-3 text-white">
-          Direct Link Has Been Copied.
-        </div>
-      )}
     </div>
   )
 }
